Listing Beers angepasst

+++ db/migrate/20141229171232_add_price_to_shop_beer.rb
@@ -1,5 +1,5 @@
 class AddPriceToShopBeer < ActiveRecord::Migration
   def change
-    add_column :shop_beers, :price, :double
+    add_column :shop_beers, :price, :float
   end
 end
